UIBULKED-179 Provide list of columns
