+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * [UIBULKED-184](https://issues.folio.org/browse/UIBULKED-184)Increase unit test code coverage for ui-bulk-edit
 * [UIBULKED-197](https://issues.folio.org/browse/UIBULKED-197) Bulk Edit app loses state when navigating to another app and back, and remembers filename but not record type
 * [UIBULKED-178](https://issues.folio.org/browse/UIBULKED-178) Displaying retrieved data
+* [UIBULKED-179](https://issues.folio.org/browse/UIBULKED-179) Provide list of columns
 
 ## [2.1.3](https://github.com/folio-org/ui-bulk-edit/tree/v2.1.3) (2022-12-02)
